Feat: 큰 수 표현 double 도입

+++ Assets/01.Scripts/Ingame/Click/ClickInfo.cs
@@ -3,10 +3,10 @@ using UnityEngine;
 public struct ClickInfo
 {
     public EClickType Type;
-    public int Damage;
+    public double Damage;
     public Vector2 Point;
 
-    public ClickInfo(EClickType type, int damage, Vector2 point = default)
+    public ClickInfo(EClickType type, double damage, Vector2 point = default)
     {
         Type = type;
         Damage = damage;
